Update modDesc.xml
Hey. Thanks for this mod. Used it in FS17. Glad I can use it now.
I translated into Russian and changed the formatting of the text a little to make it easier to translate. Translations into German and English did not change anything, only corrected a typo in the word "Sowing".
If you wish, then you can add my translation.